+++ sql-to-dbsp-compiler/SQL-compiler/src/test/java/org/dbsp/sqlCompiler/compiler/sql/functions/FunctionsTest.java
@@ -2,6 +2,11 @@ package org.dbsp.sqlCompiler.compiler.sql.functions;
 
 import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
 import org.dbsp.sqlCompiler.compiler.sql.SqlIoTest;
+import org.dbsp.sqlCompiler.compiler.sql.simple.Change;
+import org.dbsp.sqlCompiler.compiler.sql.simple.InputOutputChangeStream;
+import org.dbsp.sqlCompiler.ir.expression.DBSPTupleExpression;
+import org.dbsp.sqlCompiler.ir.expression.literal.DBSPStringLiteral;
+import org.dbsp.sqlCompiler.ir.expression.literal.DBSPZSetLiteral;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -1684,6 +1689,47 @@ public class FunctionsTest extends SqlIoTest {
                 (1 row)
                 """
         );
+    }
+
+    @Test
+    public void testGunzip() {
+        this.qs("""
+                SELECT GUNZIP(x'1f8b08000000000000ff4b4bcd49492d4a0400218115ac07000000'::bytea);
+                 gunzip
+                ------------
+                 feldera
+                (1 row)
+                
+                SELECT GUNZIP(x'1f8b08000000000000ff734bcd49492d4a0400bdb8a86307000000'::bytea);
+                 gunzip
+                ------------
+                 Feldera
+                (1 row)
+                
+                SELECT GUNZIP(x'1f8b08000000000000ffcb48cdc9c9070086a6103605000000'::bytea);
+                 gunzip
+                ------------
+                 hello
+                (1 row)
+                
+                SELECT GUNZIP(x'1f8b08000000000000132bc9c82c5600a2dc4a851282ccd48a12002e7a22ff30000000'::bytea);
+                 gunzip
+                --------------------------------------------------------------------------
+                 this is my this is my this is my this is my text
+                (1 row)
+                
+                SELECT GUNZIP(null);
+                 gunzip
+                --------
+                NULL
+                (1 row)
+                """
+        );
+    }
+
+    @Test
+    public void testGunzipRuntimeFail() {
+        this.runtimeConstantFail("SELECT GUNZIP(x'1100'::bytea)", "failed to decompress gzipped data");
     }
 
     @Test